clean up last issues
