+++ solidstart-tanstackquery-tailwind-modules/src/components/FileViewer/FileViewer.tsx
@@ -6,12 +6,12 @@ import FileText from './FileText';
 import { RepoFile } from '~/types/repo-file-type';
 import { createQuery } from '@tanstack/solid-query';
 import getRepoFile from '~/services/get-repo-file';
-import { LoadingPulseDot } from '../LoadingPulseDot/LoadingPulseDot';
 import {
   ExtensionType,
   mapExtensionToLanguage,
 } from './mapExtensionToLanguage';
 import FileCode from './FileCode';
+import { LoadingPulseDot } from '../LoadingPulseDot';
 
 const FileViewer = () => {
   const params = useParams();
