+++ solidstart-tanstackquery-tailwind-modules/src/components/LoadingPulseDot/index.ts
@@ -1,0 +1,1 @@
+export { default as LoadingPulseDot } from './LoadingPulseDot';
