+++ solidstart-tanstackquery-tailwind-modules/src/routes/[owner]/[name]/pulls.tsx
@@ -2,7 +2,7 @@ import { Switch, Match, createSignal, createEffect } from 'solid-js';
 import { createQuery } from '@tanstack/solid-query';
 import { useParams, useSearchParams } from 'solid-start';
 import getRepoInfo from '../../../services/get-repo-info';
-import { LoadingPulseDot } from '../../../components/LoadingPulseDot/LoadingPulseDot';
+import { LoadingPulseDot } from '../../../components/LoadingPulseDot';
 import { Info } from '~/types/repo-info-type';
 import { RepoHeader } from '../../../components/RepoHeader';
 import getRepoPullRequests from '../../../services/get-pull-request';
