+++ solidstart-tanstackquery-tailwind-modules/src/types/issues-type.ts
@@ -1,4 +1,4 @@
-import { Label } from "./label-type";
+import { Label } from './label-type';
 
 export interface MilestoneProps {
   id: string;
